requirements.txt
crawl4ai>=0.4.0
flask>=3.0.0
flask-login>=0.6.0
gunicorn>=22.0.0
dnspython>=2.6.0
